requirements.txt
tensorflow==1.12.2
numpy==1.16.1
scipy==1.0.0
nltk==3.4
Pillow==5.0.0
Distance==0.1.3
six==1.12.0
click==6.7
notebook>=5.7.8
tensorboard==1.12.2
tensorflow-tensorboard==0.4.0rc3